requirements.txt
fastapi
uvicorn
websockets
python-dotenv
deepgram-sdk
groq
google-generativeai
